MongoDB inclusion
The files specified under the directory name /final were an attempt to include the JSON records to the mongoDB database. The code was successfully able to create the fields according to the requirements of the user part and update records in the mongoDB which was validated inside the MongoDB compass.